requirements.txt
viam-sdk
python-socketio
requests
aiohttp